+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueAppointment.kt
@@ -219,6 +219,23 @@ data class OverdueAppointment(
     ): PagingSource<Int, OverdueAppointment>
 
     @Query("""
+      $OVERDUE_APPOINTMENTS_QUERY 
+      WHERE
+        IFNULL(patientAssignedFacilityUuid, appt_facilityUuid) = :facilityUuid AND
+        appt_scheduledDate < :scheduledBefore AND
+        appt_status != "visited"
+      GROUP BY appt_patientUuid
+      ORDER BY 
+        isAtHighRisk DESC, 
+        appt_scheduledDate DESC, 
+        appt_updatedAt ASC
+    """)
+    fun moreThanAnYearOverduePatientInFacility(
+        facilityUuid: UUID,
+        scheduledBefore: LocalDate
+    ): PagingSource<Int, OverdueAppointment>
+
+    @Query("""
       $OVERDUE_APPOINTMENTS_QUERY
       WHERE
         (